+++ georiviere/portal/tests/test_views/test_sensitivity.py
@@ -20,13 +20,11 @@ class SensitivityViewTest(TestCase):
         self.assertSetEqual(set(response.json().keys()), {'geometry', 'properties', 'type'})
         self.assertEqual(response.json()['properties']['description'], 'nuk')
 
-    def test_sensitivity_detail_json_structure(self):
+    def test_sensitivity_detail_json_content_translated(self):
         url = reverse('api_portal:sensitivities-detail',
                       kwargs={'pk': self.sensitivity.pk, 'lang': 'fr'})
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
-
-        self.assertSetEqual(set(response.json().keys()), {'id', 'name', 'description', 'attachments', 'species'})
         self.assertEqual(response.json()['description'], 'nuk')
 
     def test_sensitivity_detail_geojson_structure_en(self):
@@ -38,13 +36,13 @@ class SensitivityViewTest(TestCase):
         self.assertSetEqual(set(response.json().keys()), {'geometry', 'properties', 'type'})
         self.assertEqual(response.json()['properties']['description'], 'kun')
 
-    def test_sensitivity_detail_json_structure_en(self):
+    def test_sensitivity_detail_json_structure_translated(self):
         url = reverse('api_portal:sensitivities-detail',
                       kwargs={'pk': self.sensitivity.pk, 'lang': 'en'})
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
 
-        self.assertSetEqual(set(response.json().keys()), {'id', 'name', 'description', 'attachments', 'species'})
+        self.assertSetEqual(set(response.json().keys()), {'id', 'name', 'description', 'attachments', 'species', 'contact'})
         self.assertEqual(response.json()['description'], 'kun')
 
     def test_sensitivity_list_json_structure(self):
@@ -53,7 +51,7 @@ class SensitivityViewTest(TestCase):
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
 
-        self.assertSetEqual(set(response.json()[0].keys()), {'id', 'name', 'description', 'attachments', 'species'})
+        self.assertSetEqual(set(response.json()[0].keys()), {'id', 'name', 'description', 'attachments', 'species', 'contact'})
 
     def test_sensitivity_list_geojson_structure(self):
         url = reverse('api_portal:sensitivities-list',